Changed root directory structure (Refactor)
I made these changes because I consider necessary to keep in order which parts of the code are ready to run in the Pico and which are still under development and avoid
unexpected problems.
I also created a division between individual small function modules and complete projects that already make use of such modules giving them a hierarchy of relevance.