Update on documentation standards (Refactor)
Previously the code documentation was arbitrary and a bit verbose, a solution was implemented.
New standards were defined for line by line documentation of the code, things like date, author and so on were added.
Also reduced the overexplanation of some lines in order to keep the document cleaner and more readable.